Remove non-responsive assets
These were deprecated in 0.29.0.0.

Removed images identified like so:

    $ cat find-unused-asset
    #!/bin/bash
    BASENAME="$(basename $1)"

    if ! grep -s -E -i -r --exclude-dir ".git" --exclude-dir "tmp" --exclude-dir "log" --exclude-dir "public/assets" "$BASENAME" . > /dev/null
    then
      echo $1
    fi

    $ find app/assets/images -maxdepth 1 -type f | xargs -n1 ./find-unused-asset | xargs rm -v

This removes the image if the filename isn’t found in a file in the app.

`app/assets/images/helpmeinvestigate.png` was ound in
`doc/THEME-ASSETS-UPGRADE.md` by the script but not used in-app.

The script above removed the content type icons, as the filenames are generated
dynamically by a helper.